chore: use a single `config.rs` for testing configurations
